third_party/blink/web_tests/accessibility/presentation-owned-elements-expected.txt

These lists have a static text role because they have a presentation role.
Presentation Item 1
Presentation Item 2
The first list item has a listitem role even if ul has a presentation role because it has an explicit role.
Explicit Item 1
Implicit Item 2
These two tables have static text roles because they have presentation roles.
Presentation th
Presentation th
Presentation th
The "Presentation th" has a static text role because it has a presentation role.
Presentation th
Normal td
The row for "Explicit th" has a row role even if table has a presentation role because it has an explicit role.
Explicit th
Implicit td
The menu items except button have a static text role because it has a presentation role and is disabled.
Button has a button role because it has an explicit role.
 I have a bike   Male  Female

This test makes sure that presentation role is inherited to required owned elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

AXRole: AXGenericContainer
    AXRole: AXGenericContainer
        AXRole: AXStaticText "These lists have a static text role because they have a presentation role."
            AXRole: AXInlineTextBox "These lists have a static text role because they have a presentation role."
    AXRole: AXNone
        AXRole: AXNone "• "
            AXRole: AXStaticText "• "
                AXRole: AXInlineTextBox "• "
        AXRole: AXStaticText "Presentation Item 1"
            AXRole: AXInlineTextBox "Presentation Item 1"
    AXRole: AXNone
        AXRole: AXNone "• "
            AXRole: AXStaticText "• "
                AXRole: AXInlineTextBox "• "
        AXRole: AXStaticText "Presentation Item 2"
            AXRole: AXInlineTextBox "Presentation Item 2"
    AXRole: AXGenericContainer
        AXRole: AXStaticText "The first list item has a listitem role even if ul has a presentation role because it has an explicit role."
            AXRole: AXInlineTextBox "The first list item has a listitem role even if ul has a presentation role because it has an explicit role."
    AXRole: AXListItem
        AXRole: AXNone "• "
            AXRole: AXStaticText "• "
                AXRole: AXInlineTextBox "• "
        AXRole: AXStaticText "Explicit Item 1"
            AXRole: AXInlineTextBox "Explicit Item 1"
    AXRole: AXNone
        AXRole: AXNone "• "
            AXRole: AXStaticText "• "
                AXRole: AXInlineTextBox "• "
        AXRole: AXStaticText "Implicit Item 2"
            AXRole: AXInlineTextBox "Implicit Item 2"
    AXRole: AXGenericContainer
        AXRole: AXStaticText "These two tables have static text roles because they have presentation roles."
            AXRole: AXInlineTextBox "These two tables have static text roles because they have presentation roles."
    AXRole: AXNone
        AXRole: AXGenericContainer
            AXRole: AXGenericContainer
                AXRole: AXGenericContainer
                    AXRole: AXStaticText "Presentation th"
                        AXRole: AXInlineTextBox "Presentation th"
        AXRole: AXGenericContainer
            AXRole: AXGenericContainer
                AXRole: AXGenericContainer
                    AXRole: AXStaticText "Presentation th"
                        AXRole: AXInlineTextBox "Presentation th"
    AXRole: AXNone
        AXRole: AXGenericContainer
            AXRole: AXGenericContainer
                AXRole: AXGenericContainer
                    AXRole: AXStaticText "Presentation th"
                        AXRole: AXInlineTextBox "Presentation th"
    AXRole: AXGenericContainer
        AXRole: AXStaticText "The "Presentation th" has a static text role because it has a presentation role."
            AXRole: AXInlineTextBox "The "Presentation th" has a static text role because it has a presentation role."
    AXRole: AXTable
        AXRole: AXRowGroup
            AXRole: AXNone
                AXRole: AXGenericContainer
                    AXRole: AXStaticText "Presentation th"
                        AXRole: AXInlineTextBox "Presentation th"
            AXRole: AXRow
                AXRole: AXCell "Normal td"
                    AXRole: AXStaticText "Normal td"
                        AXRole: AXInlineTextBox "Normal td"
    AXRole: AXGenericContainer
        AXRole: AXStaticText "The row for "Explicit th" has a row role even if table has a presentation role because it has an explicit role."
            AXRole: AXInlineTextBox "The row for "Explicit th" has a row role even if table has a presentation role because it has an explicit role."
    AXRole: AXNone
        AXRole: AXGenericContainer
            AXRole: AXRow
                AXRole: AXGenericContainer
                    AXRole: AXStaticText "Explicit th"
                        AXRole: AXInlineTextBox "Explicit th"
            AXRole: AXGenericContainer
                AXRole: AXGenericContainer
                    AXRole: AXStaticText "Implicit td"
                        AXRole: AXInlineTextBox "Implicit td"
    AXRole: AXGenericContainer
        AXRole: AXStaticText "The menu items except button have a static text role because it has a presentation role and is disabled."
            AXRole: AXInlineTextBox "The menu items except button have a static text role because it has a presentation role and is disabled."
        AXRole: AXLineBreak "
"
            AXRole: AXInlineTextBox "
"
        AXRole: AXStaticText "Button has a button role because it has an explicit role."
            AXRole: AXInlineTextBox "Button has a button role because it has an explicit role."
    AXRole: AXCheckBox
    AXRole: AXStaticText " I have a bike "
        AXRole: AXInlineTextBox " I have a bike "
    AXRole: AXButton "Click me"
        AXRole: AXStaticText "Click me"
            AXRole: AXInlineTextBox "Click me"
    AXRole: AXRadioButton
    AXRole: AXStaticText " Male "
        AXRole: AXInlineTextBox " Male "
    AXRole: AXRadioButton
    AXRole: AXStaticText " Female"
        AXRole: AXInlineTextBox " Female"

PASS successfullyParsed is true

TEST COMPLETE
